+++ src/components/Miscellaneous/IconPreviewSquare/helpers/getIconPreviewBackgroundColor.ts
@@ -1,5 +1,5 @@
 import { CSSProperties } from "react";
-import { RgbValue } from "../../ColorPicker";
+import { RgbValue } from "../../../Inputs";
 
 export const getIconPreviewBackgroundColor = (rgbColor: RgbValue): CSSProperties["backgroundColor"] => {
   return `rgba(${rgbColor.r},${rgbColor.g},${rgbColor.b}, 0.25)`;