+++ src/components/Miscellaneous/IconPreviewSquare/index.ts
@@ -1,0 +1,2 @@
+export { default as IconPreviewSquare } from "./IconPreviewSquare";
+export * from "./types";